move tests to room file
