+++ tests/end-to-end/api/10-subscriptions.js
@@ -2,7 +2,6 @@ import { expect } from 'chai';
 
 import { getCredentials, api, request, credentials } from '../../data/api-data.js';
 import { createRoom } from '../../data/rooms.helper';
-import { createUser } from '../../data/users.helper';
 
 describe('[Subscriptions]', function() {
 	this.retries(0);
@@ -66,49 +65,6 @@ describe('[Subscriptions]', function() {
 		});
 	});
 
-	describe('update group dms name', () => {
-		before(async () => {
-			const testUser = await createUser();
-			const admin = 'rocketchat.internal.admin.test';
-			const rocketcat = 'rocket.cat';
-
-			console.log('test user: ', testUser.username);
-			const usernames = [testUser.username, admin, rocketcat].join(',');
-
-			request.post(api('dm.create'))
-				.set(credentials)
-				.send({
-					usernames,
-				})
-				.end((err, res) => {
-					this.roomId = res.body.room.rid;
-					this.testUser = testUser;
-				});
-		});
-
-		it('should update group name', (done) => {
-			setTimeout(() => {
-				request.post(api('users.update'))
-					.set(credentials)
-					.send({
-						userId: this.testUser._id,
-						data: {
-							username: `changed.${ this.testUser.username }`,
-						},
-					})
-					.end(() => {
-					  request.get(api('rooms.info'))
-					  .set(credentials)
-					  .query({ roomId: this.roomId })
-					  .end((err, res) => {
-					  	const { room } = res.body;
-					  	expect(room.usernames.includes(`changed.${ this.testUser.username }`)).to.be.true;
-					  	done();
-				  	});
-				  });
-			}, 200);
-		});
-	});
 
 	describe('[/subscriptions.read]', () => {
 		let testChannel;
